--- a/app/config.go
+++ b/app/config.go
@@ -17,12 +17,10 @@ import (
 
 	kitlevel "github.com/go-kit/kit/log/level"
 	"github.com/go-kit/kit/log/term"
-	"github.com/pokt-network/pocket-core/baseapp"
 	"github.com/pokt-network/pocket-core/codec"
 	types2 "github.com/pokt-network/pocket-core/codec/types"
 	"github.com/pokt-network/pocket-core/crypto"
 	kb "github.com/pokt-network/pocket-core/crypto/keys"
-	"github.com/pokt-network/pocket-core/store"
 	sdk "github.com/pokt-network/pocket-core/types"
 	"github.com/pokt-network/pocket-core/types/module"
 	apps "github.com/pokt-network/pocket-core/x/apps"
@@ -222,6 +220,11 @@ func UpdateConfig(datadir string) {
 
 	//Write Defaults on GlobalConfig.
 	GlobalConfig.TendermintConfig.LevelDBOptions = config2.DefaultLevelDBOpts()
+	GlobalConfig.TendermintConfig.LevelDBOptions.BlockSize = 1024
+	GlobalConfig.TendermintConfig.LevelDBOptions.BlockCacheCapacity = 32 * 1024 * 1024
+	GlobalConfig.TendermintConfig.LevelDBOptions.DisableBufferPool = false
+	GlobalConfig.TendermintConfig.LevelDBOptions.OpenFilesCacheCapacity = 1000
+	GlobalConfig.TendermintConfig.LevelDBOptions.WriteBuffer = 128 * 1024 * 1024
 	sdk.DefaultPocketConsensusConfig(GlobalConfig.TendermintConfig.Consensus)
 	GlobalConfig.TendermintConfig.P2P.AllowDuplicateIP = true
 	GlobalConfig.TendermintConfig.P2P.AddrBookStrict = false
@@ -360,7 +363,7 @@ func InitTendermint(keybase bool, chains *types.HostedBlockchains, logger log.Lo
 		keys = MustGetKeybase()
 	}
 	appCreatorFunc := func(logger log.Logger, db dbm.DB, _ io.Writer) *PocketCoreApp {
-		return NewPocketCoreApp(nil, keys, getTMClient(), chains, logger, db, GlobalConfig.PocketConfig.Cache, GlobalConfig.PocketConfig.IavlCacheSize, baseapp.SetPruning(store.PruneNothing))
+		return NewPocketCoreApp(nil, keys, getTMClient(), chains, logger, db)
 	}
 	tmNode, app, err := NewClient(config(c), appCreatorFunc)
 	if err != nil {