--- a/app/pocket.go
+++ b/app/pocket.go
@@ -34,8 +34,7 @@ type PocketCoreApp struct {
 	// the codec (uses amino)
 	cdc *codec.Codec
 	// Keys to access the substores
-	Keys  map[string]*sdk.KVStoreKey
-	Tkeys map[string]*sdk.TransientStoreKey
+	Keys map[string]*sdk.KVStoreKey
 	// Keepers for each module
 	accountKeeper auth.Keeper
 	appsKeeper    appsKeeper.Keeper
@@ -47,24 +46,21 @@ type PocketCoreApp struct {
 }
 
 // new pocket core base
-func NewPocketBaseApp(logger log.Logger, db db.DB, cache bool, iavlCacheSize int64, options ...func(*bam.BaseApp)) *PocketCoreApp {
+func NewPocketBaseApp(logger log.Logger, db db.DB) *PocketCoreApp {
 	cdc = Codec()
 	bam.SetABCILogging(GlobalConfig.PocketConfig.ABCILogging)
 	// BaseApp handles interactions with Tendermint through the ABCI protocol
-	bApp := bam.NewBaseApp(appName, logger, db, cache, iavlCacheSize, auth.DefaultTxDecoder(cdc), cdc, options...)
+	bApp := bam.NewBaseApp(appName, logger, db, auth.DefaultTxDecoder(cdc), cdc)
 	// set version of the baseapp
 	bApp.SetAppVersion(AppVersion)
 	// setup the key value store Keys
 	k := sdk.NewKVStoreKeys(bam.MainStoreKey, auth.StoreKey, nodesTypes.StoreKey, appsTypes.StoreKey, gov.StoreKey, pocketTypes.StoreKey)
-	// setup the transient store Keys
-	tkeys := sdk.NewTransientStoreKeys(nodesTypes.TStoreKey, appsTypes.TStoreKey, pocketTypes.TStoreKey, gov.TStoreKey)
 	// add params Keys too
 	// Create the application
 	return &PocketCoreApp{
 		BaseApp: bApp,
 		cdc:     cdc,
 		Keys:    k,
-		Tkeys:   tkeys,
 	}
 }