--- a/go.mod
+++ b/go.mod
@@ -19,7 +19,6 @@ require (
 	github.com/tendermint/go-amino v0.15.1
 	github.com/tendermint/tendermint v0.33.7
 	github.com/tendermint/tm-db v0.5.1
-	github.com/tidwall/btree v1.3.1
 	github.com/willf/bloom v2.0.3+incompatible
 	golang.org/x/crypto v0.0.0-20210915214749-c084706c2272
 	google.golang.org/protobuf v1.27.1