--- a/proto/protocgen.sh
+++ b/proto/protocgen.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 
+cd ..
+
 set -eo pipefail
 
-proto_dirs=$(find ./proto -path -prune -o -name '*.proto' -print0 | xargs -0 -n1 dirname | sort | uniq)
+proto_dirs=$(find proto -path -prune -o -name '*.proto' -print0 | xargs -0 -n1 dirname | sort | uniq)
 for dir in $proto_dirs; do
   protoc \
   -I "proto" \