--- a/store/iavl/wire.go
+++ b/store/iavl/wire.go
@@ -1,8 +1,0 @@
-package iavl
-
-import (
-	"github.com/pokt-network/pocket-core/codec"
-	"github.com/pokt-network/pocket-core/codec/types"
-)
-
-var cdc = codec.NewCodec(types.NewInterfaceRegistry())